chore: restructure core modules (views, entity, feature)
