+++ sdk/python/feast/core/__init__.py
@@ -1,0 +1,2 @@
+from .entity import Entity
+from .feature import Feature
