+++ sdk/python/feast/core/feature_views/__init__.py
@@ -1,0 +1,5 @@
+from .batch_view import BatchFeatureView
+from .generic import FeatureView
+from .on_demand_view import OnDemandFeatureView, on_demand_feature_view
+from .request_view import RequestFeatureView
+from .stream_view import StreamFeatureView, stream_feature_view
